Renamed SomePointers.js to LeafletLocalData
