+++ src/client/components/LeafletLocalData.js
